docs : handlers.ts -> musicbarHandlers, genreHandlers로 나눔
